fcuk

+++ bot.py
@@ -204,17 +204,18 @@ async def on_message(message):
     # Portal
     match = await bot.db.portal.find_one({"id": message.guild.id})
     # All portal channels to send to EXCEPT the current one
-    channels = [x['channel'] for x in (await bot.db.portal.find().to_list(None))].remove(match["channel"]) 
-    
-    img = message.author.avatar_url_as(format="png", size=1024)
-    byte = await (await bot.session.get(img)).read()
-    for chan in channels:
-        current = bot.get_channel(chan)
-        if current:
-            webhook = await current.create_webhook(name=message.author.display_name, avatar=byte)
-            await webhook.send(message.content.replace("@", "@\u200b"))
-            await webhook.delete()
-    #if message.channel.id == 566030691360440356 and not message.author.discriminator == "0000": # DBBBH Side
+    if match:
+        channels = [x['channel'] for x in (await bot.db.portal.find().to_list(None))].remove(match["channel"]) 
+        
+        img = message.author.avatar_url_as(format="png", size=1024)
+        byte = await (await bot.session.get(img)).read()
+        for chan in channels:
+            current = bot.get_channel(chan)
+            if current:
+                webhook = await current.create_webhook(name=message.author.display_name, avatar=byte)
+                await webhook.send(message.content.replace("@", "@\u200b"))
+                await webhook.delete()
+        #if message.channel.id == 566030691360440356 and not message.author.discriminator == "0000": # DBBBH Side
     #    img = message.author.avatar_url_as(format="png", size=1024)
     #    byte = await (await bot.session.get(img)).read()
     #    webhook = await bot.get_channel(566015775261982866).create_webhook(name=message.author.display_name, avatar=byte)
